prep for new command structure
