+++ classes/sql/modules/SqlCommands.js
@@ -1,4 +1,5 @@
 "use strict"
+const sqlPool = require('../Sql').pool
 const util = require('util')
 
 module.exports = class SqlCommands {
@@ -26,5 +27,40 @@ module.exports = class SqlCommands {
     }
 
     return returnData
+  }
+
+  static async getCommandData () {
+    let results = await sqlPool.query(`
+        SELECT c.ID as 'commandID', 
+               cG.ID as 'commandGroupID', 
+               cG.name as 'commandGroupName', 
+               cGL.channelID as 'channelID', 
+               cGL.botID as 'botID',
+               command, 
+               response, 
+               userLevel, 
+               cooldown, 
+               timesUsed
+        FROM commands c, commandGroup cG, commandGroupLink cGL 
+        WHERE c.isRegex = b'0'
+          AND c.enabled = b'1'
+          AND cG.enabled = b'1'
+          AND cGL.enabled = b'1' 
+          AND c.commandGroupID = cG.ID
+        ;`)
+
+    let resultsRegex = await sqlPool.query(`
+        SELECT ID, command, response, userLevel, cooldown, timesUsed
+        FROM globalCommands
+        WHERE enabled = b'1'
+          AND isRegex = b'1';`)
+
+    return SqlCommands.resultDataFromResults(results, resultsRegex)
+  }
+
+  static increaseTimesUsed (commandID) {
+    sqlPool.query(`UPDATE globalCommands 
+      set timesUsed = timesUsed + 1 
+      WHERE ID = ?;`, commandID)
   }
 }
